Fixed online phase

+++ src/common/online.h
@@ -61,7 +61,7 @@ void print_hash(const unsigned char *digest);
  * @param mt the number of endpoints in the list.
  * @return the index of found value in the list if found, -1 otherwise.
  */
-int search_endpoint(char **endpoints, char *plain_text, int mt, int pwd_length);
+unsigned long search_endpoint(char **endpoints, char *plain_text, int mt, int pwd_length);
 
 /**
  * Transforms a char array to a password.
